// qa-framework/src/library/models/ParameterType.ts
enum ParameterType {
  Title = 'Title',
  Text = 'Text',
  XPath = 'XPath',
  CSS = 'CSS',
  Attribute = 'Attribute',
  Enter = 'Enter',
  Tab = 'Tab',
  Milliseconds = 'MilliSeconds',
  Seconds = 'Seconds',
  URL = 'URL',
  Dropdown = 'Dropdown',
  Clear = 'Clear',
  AssertContains = 'AssertContains', 
  AssertStartsWith = 'AssertStartsWith',
  AssertEndsWith = 'AssertEndsWith',
  AssertGreaterThan = 'AssertGreaterThan',
  AssertLessThan = 'AssertLessThan'
}

export default ParameterType;